refactor: Move validation result factory up for reuse
