+++ backend/check_stac_metadata/utils.py
@@ -2,7 +2,7 @@ from argparse import ArgumentParser, Namespace
 from json import dumps, load
 from logging import Logger
 from os.path import dirname, join
-from typing import Callable, Dict, List, Optional
+from typing import Callable, Dict, List
 
 from botocore.response import StreamingBody  # type: ignore[import]
 from jsonschema import (  # type: ignore[import]
@@ -13,10 +13,9 @@ from jsonschema import (  # type: ignore[import]
 )
 from jsonschema._utils import URIDict  # type: ignore[import]
 
-from ..check import Check
 from ..processing_assets_model import ProcessingAssetType, ProcessingAssetsModel
 from ..types import JsonObject
-from ..validation_results_model import ValidationResult, ValidationResultsModel
+from ..validation_results_model import ValidationResult, ValidationResultFactory
 
 
 class STACSchemaValidator(Draft7Validator):
@@ -45,21 +44,6 @@ class STACSchemaValidator(Draft7Validator):
         with open(join(self.script_dir, path)) as file_pointer:
             schema_dict: JsonObject = load(file_pointer)
             return schema_dict
-
-
-class ValidationResultFactory:  # pylint:disable=too-few-public-methods
-    def __init__(self, hash_key: str):
-        self.hash_key = hash_key
-
-    def save(
-        self, url: str, result: ValidationResult, details: Optional[JsonObject] = None
-    ) -> None:
-        ValidationResultsModel(
-            pk=self.hash_key,
-            sk=f"CHECK#{Check.JSON_SCHEMA.value}#URL#{url}",
-            result=result.value,
-            details=details,
-        ).save()
 
 
 class STACDatasetValidator:
